% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cloudbandmath}
\alias{cloudbandmath}
\title{This function contains formulas for calculating Vegatation indices from Landsat-8}
\usage{
cloudbandmath(multispectral.image)
}
\arguments{
\item{multispectral.image}{A ee.Image object from which calculations will be made}
}
\description{
This function contains formulas for calculating Vegatation indices from Landsat-8
}
\keyword{internal}
