% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datesFromNames}
\alias{datesFromNames}
\title{Gets Dates for Landsat Scene Names}
\usage{
datesFromNames(input)
}
\arguments{
\item{input}{Either a data.frame or a list containing Rasters.}
}
\description{
Gets Dates for Landsat Scene Names
}
\keyword{internal}
