% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senseLandsat.R
\name{senseLandsat}
\alias{senseLandsat}
\title{Shapefile to Earth Engine Polygon}
\usage{
senseLandsat(ee.geometry)
}
\arguments{
\item{ee.geometry}{Google Earth Engine Object ee.Geometry.Polygon}
}
\value{
a data.frame containing all band values and VIs.
}
\description{
This function receives an Earth Engine Polygon, gets Landsat-8
multispectral images and calculates VIs for that location
}
\examples{
ee.geometry <- shpToEE(shapefile="D:/Dropbox/Science/reSense/data/hudson.shp")
hudson.df <- senseLandsat(ee.geometry)
}
