% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shptoEE.R
\name{shpToEE}
\alias{shpToEE}
\title{Shapefile to Earth Engine Polygon}
\usage{
shpToEE(shapefile)
}
\arguments{
\item{shapefile}{Path to a .shp file}
}
\value{
Google Earth Engine Polygon
}
\description{
This function converts a Shapefile (.shp) to a
Google Earth Engine Polygon (ee.Geometry.Polygon)
}
\examples{
ee.geometry <- shpToEE(shapefile="D:/Dropbox/Science/reSense/data/hudson.shp")


}
